#1 - Improved `#ensureConnection` Sync w/ alphanum. compare
- [x] Avoid double-connecting peers (both send outgoing at same time) by using a deterministic comparison
- [x] Fixed stale friend data overwriting newer member data
- [x] Fixed re-emitting of handled events by deferring handler functions
- [x] Added JSDoc code comments
- [x] Split `Friend` file for multi-file reuse
