+++ src/Friend.ts
@@ -1,0 +1,13 @@
+import Peer from "peerjs";
+
+/** Metadata regarding a member of the party */
+export interface Friend {
+    id: string;
+    muted: boolean;
+    title: string;
+}
+
+/** An active connection to a known member of the party */
+export interface FriendConnected extends Friend {
+    conn: Peer.DataConnection;
+}
